Fix: Allow selecting the "Choose" option in the origin question
This commit fixes an issue where the "Choose" option in the origin question could not be selected, forcing the user to select another option.

The issue was caused by the `chosenIndex` being set to null when the user selected the "Choose" option, which resulted in the option being immediately deselected.

This commit fixes the issue by allowing the `chosenIndex` to be set to 0, which represents the "Choose" option. This allows the user to select and keep the "Choose" option selected.