Refactor: Move note widget to separate function
This commit refactors the handling of the note widget by moving it to a separate function `getNote` in the `Site` interface.

Previously, the note widget was included in the list of widgets returned by `getWidgets`. This change separates it to allow for more specific handling and placement within the UI.

Additionally, the commit removes the unnecessary sending dialog during questionnaire submission.